# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type LeadDetails = { [key: string]: unknown } | null;

export interface Lead {
  id: number;
  source: string;
  name: string;
  email: string;
  phone?: string | null;
  businessName?: string | null;
  message?: string | null;
  details?: LeadDetails;
  createdAt: string;
}

export type LeadCreateSource = typeof LeadCreateSource[keyof typeof LeadCreateSource];


export const LeadCreateSource = {
  'web-demo': 'web-demo',
  'web-contact': 'web-contact',
  'mobile-demo': 'mobile-demo',
  'mobile-contact': 'mobile-contact',
} as const;

export type LeadCreateDetails = { [key: string]: unknown };

export interface LeadCreate {
  source: LeadCreateSource;
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  phone?: string;
  businessName?: string;
  message?: string;
  details?: LeadCreateDetails;
  /** Honeypot field — must be left empty. Filled values mark the submission as bot traffic. */
  website?: string;
}

export interface LeadCreated {
  id: number;
  status: string;
}

export interface ApiError {
  message: string;
}

export interface HealthStatus {
  status: string;
}

export type ListLeads200 = {
  leads: Lead[];
};