# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadDetails } from './leadDetails';

export interface Lead {
  id: number;
  source: string;
  name: string;
  email: string;
  phone?: string | null;
  businessName?: string | null;
  message?: string | null;
  details?: LeadDetails;
  createdAt: Date;
}