# lib/api-zod/src/generated/types/leadCreate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadCreateDetails } from './leadCreateDetails';
import type { LeadCreateSource } from './leadCreateSource';

export interface LeadCreate {
  source: LeadCreateSource;
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  phone?: string;
  businessName?: string;
  message?: string;
  details?: LeadCreateDetails;
  /** Honeypot field — must be left empty. Filled values mark the submission as bot traffic. */
  website?: string;
}